requirements.txt
google-generativeai
pydantic
rich
psutil
tenacity
requests
python-dotenv
